refactor(profile): rename components
